+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Additional handler options to error handler
 - Conditional options
 
+### Changed
+
+- Renamed `ApplicationOptions` to `Options`
+
 
 ## 0.1.0 - 2017-10-01